Project folder restructure
