+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM alpine:latest
+FROM golang:1.23.1-alpine3.20
 
 # Add the commands needed to put your compiled go binary in the container and
 # run it when the container starts.
@@ -11,3 +11,18 @@ FROM alpine:latest
 # "kadlab", which you do by using the following command:
 #
 # $ docker build . -t kadlab
+
+# Set working directory
+WORKDIR /app
+
+# Copy the source code into the container
+COPY . .
+
+# Install dependencies
+RUN go mod download
+
+# Build the binary
+RUN go build -o kadlab ./cmd/kademlia_main
+
+# Run the binary
+CMD ["./kadlab"]
