+++ cmd/kademlia_main/main.go
@@ -3,15 +3,15 @@
 package main
 
 import (
-	"kadlab-group-6/kademlia"
 	"fmt"
+	node "kadlab-group-6/pkg/kademlia_node"
 )
 
 func main() {
 	fmt.Println("Pretending to run the kademlia app...")
 	// Using stuff from the kademlia package here. Something like...
-	id := kademlia.NewKademliaID("FFFFFFFF00000000000000000000000000000000")
-	contact := kademlia.NewContact(id, "localhost:8000")
+	id := node.NewKademliaID("FFFFFFFF00000000000000000000000000000000")
+	contact := node.NewContact(id, "localhost:8000")
 	fmt.Println(contact.String())
 	fmt.Printf("%v\n", contact)
 }
