+++ pkg/kademlia_node/contact.go
@@ -1,4 +1,4 @@
-package kademlia
+package kademlia_node
 
 import (
 	"fmt"
