+++ pkg/kademlia_node/network.go
@@ -1,4 +1,5 @@
-package kademlia
+package kademlia_node
+
 
 type Network struct {
 }
